add config check mechanism
