+++ server/server.go
@@ -14,18 +14,21 @@
 package server
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"math/rand"
 	"net/http"
 	"path"
 	"path/filepath"
+	"reflect"
 	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
+	"github.com/BurntSushi/toml"
 	"github.com/coreos/go-semver/semver"
 	"github.com/golang/protobuf/proto"
 	"github.com/gorilla/mux"
@@ -38,6 +41,7 @@ import (
 	"github.com/pingcap/kvproto/pkg/pdpb"
 	"github.com/pingcap/log"
 	"github.com/pingcap/pd/pkg/dashboard/uiserver"
+	pd "github.com/pingcap/pd/client"
 	"github.com/pingcap/pd/pkg/etcdutil"
 	"github.com/pingcap/pd/pkg/grpcutil"
 	"github.com/pingcap/pd/pkg/logutil"
@@ -45,7 +49,7 @@ import (
 	"github.com/pingcap/pd/pkg/ui"
 	"github.com/pingcap/pd/server/cluster"
 	"github.com/pingcap/pd/server/config"
-	configmanager "github.com/pingcap/pd/server/config_manager"
+	"github.com/pingcap/pd/server/config_manager"
 	"github.com/pingcap/pd/server/core"
 	"github.com/pingcap/pd/server/id"
 	"github.com/pingcap/pd/server/kv"
@@ -81,6 +85,12 @@ var (
 	EnableZap = false
 	// EtcdStartTimeout the timeout of the startup etcd.
 	EtcdStartTimeout = time.Minute * 5
+)
+
+const (
+	// Component is used to represent the kind of component in config manager.
+	Component           = "pd"
+	configCheckInterval = 1 * time.Second
 )
 
 // Server is the pd server.
@@ -127,6 +137,9 @@ type Server struct {
 
 	// components' configuration management
 	cfgManager *configmanager.ConfigManager
+	// component config
+	configVersion *configpb.Version
+	configClient  pd.ConfigClient
 }
 
 // HandlerBuilder builds a server HTTP handler.
@@ -434,7 +447,10 @@ func (s *Server) startServerLoop(ctx context.Context) {
 	go s.leaderLoop()
 	go s.etcdLeaderLoop()
 	go s.serverMetricsLoop()
-
+	if s.cfg.EnableConfigManager {
+		s.serverLoopWg.Add(1)
+		go s.configCheckLoop()
+	}
 }
 
 func (s *Server) stopServerLoop() {
@@ -858,6 +874,18 @@ func (s *Server) SetLogLevel(level string) {
 	log.Warn("log level changed", zap.String("level", log.GetLevel().String()))
 }
 
+// GetConfigVersion returns the config version.
+func (s *Server) GetConfigVersion() *configpb.Version {
+	if s.configVersion == nil {
+		return &configpb.Version{Local: 0, Global: 0}
+	}
+	return s.configVersion
+}
+
+func (s *Server) setConfigVersion(version *configpb.Version) {
+	s.configVersion = version
+}
+
 func (s *Server) leaderLoop() {
 	defer logutil.LogPanic()
 	defer s.serverLoopWg.Done()
@@ -991,6 +1019,126 @@ func (s *Server) etcdLeaderLoop() {
 			return
 		}
 	}
+}
+
+func (s *Server) configCheckLoop() {
+	defer logutil.LogPanic()
+	defer s.serverLoopWg.Done()
+
+	ctx, cancel := context.WithCancel(s.serverLoopCtx)
+	defer cancel()
+
+	ticker := time.NewTicker(configCheckInterval)
+	defer ticker.Stop()
+	for {
+		// wait leader
+		leader := s.GetLeader()
+		if leader != nil {
+			var err error
+			securityConfig := s.GetSecurityConfig()
+			s.configClient, err = pd.NewConfigClientWithContext(ctx, s.GetEndpoints(), pd.SecurityOption{
+				CAPath:   securityConfig.CAPath,
+				CertPath: securityConfig.CertPath,
+				KeyPath:  securityConfig.KeyPath,
+			})
+			if err != nil {
+				log.Error("failed to create config client", zap.Error(err))
+				return
+			}
+			break
+		}
+		select {
+		case <-ticker.C:
+		case <-ctx.Done():
+			log.Info("config check stop running")
+			return
+		}
+	}
+
+	addr := s.GetAddr()
+	version := s.GetConfigVersion()
+	var config bytes.Buffer
+	if err := toml.NewEncoder(&config).Encode(*s.GetConfig()); err != nil {
+		log.Error("failed to encode config", zap.Error(err))
+		return
+	}
+	if err := s.createComponentConfig(ctx, version, addr, config.String()); err != nil {
+		log.Error("failed to create config", zap.Error(err))
+		return
+	}
+
+	for {
+		select {
+		case <-ctx.Done():
+			log.Info("config check has been stopped")
+			return
+		case <-ticker.C:
+			version := s.GetConfigVersion()
+			config, err := s.getComponentConfig(ctx, version, addr)
+			if err != nil {
+				log.Error("failed to get config", zap.Error(err))
+			}
+			if config == "" {
+				continue
+			}
+			if err := s.updateComponentConfig(config); err != nil {
+				log.Error("failed to update config", zap.Error(err))
+			}
+		}
+	}
+}
+
+func (s *Server) createComponentConfig(ctx context.Context, version *configpb.Version, componentID, config string) error {
+	status, v, config, err := s.configClient.Create(ctx, version, Component, componentID, config)
+	if err != nil {
+		return err
+	}
+	switch status.GetCode() {
+	case configpb.StatusCode_OK, configpb.StatusCode_WRONG_VERSION:
+		s.setConfigVersion(v)
+		s.updateComponentConfig(config)
+	case configpb.StatusCode_UNKNOWN:
+		return errors.Errorf("unknown error: %v", status.GetMessage())
+	}
+	return nil
+}
+
+func (s *Server) getComponentConfig(ctx context.Context, version *configpb.Version, componentID string) (string, error) {
+	status, v, cfg, err := s.configClient.Get(ctx, version, Component, componentID)
+	if err != nil {
+		return "", err
+	}
+	var config string
+	switch status.GetCode() {
+	case configpb.StatusCode_OK:
+		config = cfg
+	case configpb.StatusCode_WRONG_VERSION:
+		s.setConfigVersion(v)
+	case configpb.StatusCode_UNKNOWN:
+		return "", errors.Errorf("unknown error: %v", status.GetMessage())
+	}
+	return config, nil
+}
+
+func (s *Server) updateComponentConfig(cfg string) error {
+	old := s.GetConfig()
+	new := &config.Config{}
+	if _, err := toml.Decode(cfg, &new); err != nil {
+		return err
+	}
+	var err error
+	if !reflect.DeepEqual(old.Schedule, new.Schedule) {
+		err = s.SetScheduleConfig(new.Schedule)
+	}
+
+	if !reflect.DeepEqual(old.Replication, new.Replication) {
+		err = s.SetReplicationConfig(new.Replication)
+	}
+
+	if !reflect.DeepEqual(old.PDServerCfg, new.PDServerCfg) {
+		err = s.SetPDServerConfig(new.PDServerCfg)
+	}
+	return err
 }
 
 func (s *Server) reloadConfigFromKV() error {
